+++ changelog.md
@@ -1,5 +1,9 @@
 GeekPark.js更新日志
 --------
+###1.1.1 
+1. 增加loadTemplate函数，用于动态加载HTML Template
+2. 增加isPage，通过body的classname来判断js是否应该运行
+
 ###1.1.0 更新变动
 1. 删除1.0.0版本中的检测CSS Transition执行完毕
 2. 删除1.0.0版本中的微信分享自定义，因为微信官方已屏蔽此接口